+++ Deployment/config.json
@@ -1,0 +1,4 @@
+{
+  "clientId": "552272683543560194",
+  "guildId": "537484725896478733"
+}
