--- SlashCommands/slashdata.js
@@ -1,7 +1,0 @@
-
-/**@type {import("discord.js").ApplicationCommandData} */
-const SlashCommands = {
-
-}
-
-module.exports = { SlashCommands };
