--- commands/private/interactionUpdater.js
@@ -1,15 +1,0 @@
-const { Command } = require('discord-akairo');
-
-class InteractionUpdater extends Command {
-	constructor() {
-		super('updateInteraction', {
-			aliases: ['updateInteraction'],
-		});
-	}
-
-	async exec(message, args) {
-
-	}
-}
-
-module.exports = InteractionUpdater;
